Updated Lookup to KvpC with updates to KvpH and KvpS
Consolidating naming of Keyed collections so we can introduce Ikvp interface
Upgraded Lookup to KvpC.  This meant that sections of KvpH and KvpS were also updated  to improve code.

Some follow up correction for renaming of enumerations to use a tw letter prefix with double underscore for enum type and single underscore for enum members.